updated slate version
